requirements.txt
discord.py
janome